refactor filenames for easier organization
