+++ app/routers/docs_router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, UploadFile
 from fastapi.responses import FileResponse
 
-from ..services import docs
+from ..services import assets_service
 
 router = APIRouter(
     prefix="/assets",
@@ -13,25 +13,25 @@ async def upload_files(file: UploadFile):
     """
     Upload a file and return its filename.
     """
-    return await docs.save_file(file)
+    return await assets_service.save_file(file)
 
 @router.get("/")
 async def get_docs():
     """
     Retrieve all documents.
     """
-    return await docs.get_all_documents()
+    return await assets_service.get_all_documents()
 
 @router.get("/search")
 async def search_docs(query: str):
     """
     Search documents in the vector database.
     """
-    return await docs.search_documents(query)
+    return await assets_service.search_documents(query)
 
 @router.get("/debug/search")
 async def debug_search_docs(query: str):
     """
     Debug search documents in the vector database.
     """
-    return await docs.debug_search_documents(query)
+    return await assets_service.debug_search_documents(query)
