+++ app/services/tasks_service.py
@@ -1,4 +1,4 @@
-from ..repositories.primary import tasks as task_repo
+from ..repositories.primary import tasks_repo as task_repo
 from ..models.task import Task
 
 async def get_all_tasks() -> list[Task]:
