+++ app/utils/fs_utils.py
@@ -1,6 +1,6 @@
 import os
-from .env import get_config
-from ..constants import env as env_constants
+from .env_utils import get_config
+from ..constants import env_constants as env_constants
 
 def check_or_create_directory(dir_path):
     """Check if a directory exists, and create it if it doesn't."""
